+++ OnRail/Extensions/Try/TryExtensions.cs
@@ -11,8 +11,7 @@ public static partial class TryExtensions {
 
         for (var counter = 0; counter < numOfTry; counter++) {
             try {
-                return Result<T>.Ok(function())
-                    .AddNumOfTry(counter + 1, numOfTry);
+                return Result<T>.Ok(function());
             }
             catch (Exception e) {
                 errors.Add(e);
@@ -39,8 +38,7 @@ public static partial class TryExtensions {
                 var result = function();
 
                 if (result.IsSuccess || numOfTry == 1) {
-                    return result
-                        .AddNumOfTry(counter + 1, numOfTry);
+                    return result;
                 }
 
                 if (result.Detail is not null)
@@ -67,8 +65,7 @@ public static partial class TryExtensions {
                 var result = function();
 
                 if (result.IsSuccess || numOfTry == 1) {
-                    return result
-                        .AddNumOfTry(counter + 1, numOfTry);
+                    return result;
                 }
 
                 if (result.Detail is not null)
@@ -92,7 +89,7 @@ public static partial class TryExtensions {
         for (var counter = 0; counter < numOfTry; counter++) {
             try {
                 action();
-                return Result.Ok().AddNumOfTry(counter + 1, numOfTry);
+                return Result.Ok();
             }
             catch (Exception e) {
                 errors.Add(e);